requirements.txt
discord.py==2.2.2
PyNaCl==1.5.0
python-dotenv==0.21.0    
youtube-dl==2021.12.17 
requests==2.28.1
DiscordUtils==1.3.4
numpy==1.24.3